[MERGE] Revisão de código.

+++ flow/core/src/main/java/br/net/mirante/singular/flow/core/ProcessDataServiceImpl.java
@@ -35,7 +35,7 @@ public class ProcessDataServiceImpl<I extends ProcessInstance> implements IProce
     }
 
     @Override
-    public final I retrieveInstance(Integer entityCod) {
+    public final I retrieveInstance(Long entityCod) {
         IEntityProcessInstance entityProcessInstance = getPersistenceService().retrieveProcessInstanceByCod(entityCod);
         if (entityProcessInstance != null) {
             return processDefinition.convertToProcessInstance(entityProcessInstance);
